requirements.txt
azure-mgmt-containerregistry
azure-identity
requests